refactor: better component name (FileViewer -> IntentHandler)
